src/utils/theme_family.rs
```rust
use std::collections::BTreeMap;
use std::borrow::Cow;
use super::theme::Theme;
use super::color_scale_set::ColorScaleSet;

/// Represents a family of themes.
/// A ThemeFamily represents the theme as a whole i.e., a name, author, optionally, scales and overrides, plus a vector of Themes.
/// A ThemeFamily with defined scales need not have any Themes.
#[derive(Debug)]
pub struct ThemeFamily<'a> {
    name: Cow<'a, str>,
    author: Cow<'a, str>,
    scales: Option<BTreeMap<Cow<'a, str>, ColorScaleSet>>,
    overrides: Option<BTreeMap<Cow<'a, str>, Cow<'a, str>>>,
    themes: Vec<Theme<'a>>,
}

impl<'a> ThemeFamily<'a> {
    /// Creates a new ThemeFamily.
    pub fn new(name: Cow<'a, str>, author: Cow<'a, str>) -> Self {
        Self {
            name,
            author,
            scales: None,
            overrides: None,
            themes: Vec::new(),
        }
    }

    /// Sets the scales for the ThemeFamily.
    pub fn set_scales(&mut self, scales: BTreeMap<Cow<'a, str>, ColorScaleSet>) {
        self.scales = Some(scales);
    }

    /// Sets the overrides for the ThemeFamily.
    pub fn set_overrides(&mut self, overrides: BTreeMap<Cow<'a, str>, Cow<'a, str>>) {
        self.overrides = Some(overrides);
    }

    /// Adds a theme to the ThemeFamily.
    pub fn add_theme(&mut self, theme: Theme<'a>) {
        self.themes.push(theme);
    }
}
```